+++ plasmapy/particles/particle_collections.py
@@ -436,60 +436,6 @@ Remove the first occurrence of a
 
 ParticleList.reverse.__doc__ = """Reverse the |ParticleList| in place."""
 
-
-@particle_input(any_of={"element", "isotope", "ion"})
-def ionic_levels(
-    particle: Particle,
-    min_charge: Integral = 0,
-    max_charge: Optional[Integral] = None,
-) -> ParticleList:
-    """
-    Return a |ParticleList| that includes different ionic levels of a
-    base atom.
-
-    Parameters
-    ----------
-    particle : `~plasmapy.particles.particle_class.ParticleLike`
-        Representation of an element, ion, or isotope.
-
-    min_charge : integer, optional
-        The starting charge number. Defaults to ``0``.
-
-    max_charge : integer, optional
-        The ending charge number, which will be included in the
-        |ParticleList|.  Defaults to the atomic number.
-
-    Returns
-    -------
-    `~plasmapy.particles.particle_collections.ParticleList`
-        The ionic levels of the atom provided from ``min_charge`` to
-        ``max_charge``.
-
-    Examples
-    --------
-    >>> from plasmapy.particles import ionic_levels
-    >>> ionic_levels("He")
-    ParticleList(['He 0+', 'He 1+', 'He 2+'])
-    >>> ionic_levels("Fe-56", min_charge=13, max_charge=15)
-    ParticleList(['Fe-56 13+', 'Fe-56 14+', 'Fe-56 15+'])
-    """
-    base_particle = Particle(particle.isotope or particle.element)
-
-    if max_charge is None:
-        max_charge = particle.atomic_number
-
-    if not min_charge <= max_charge <= particle.atomic_number:
-        raise ChargeError(
-            f"Need min_charge ({min_charge}) "
-            f"≤ max_charge ({max_charge}) "
-            f"≤ atomic number ({base_particle.atomic_number})."
-        )
-
-    return ParticleList(
-        [Particle(base_particle, Z=Z) for Z in range(min_charge, max_charge + 1)]
-    )
-
-
 ParticleListLike = Union[ParticleList, Sequence[ParticleLike]]
 
 ParticleListLike.__doc__ = r"""
